feat: add path of raw json data file as output
Now it's possible to get the raw json data from the pull request as used
in the action itself. This can be handy if someone wants to collect this
data in a different place like a database over multiple repos.

With the path of the raw json data file he can upload it to a central
database for further processing.