fix: escape markdown/html comments in json fields
There is a problem to create pretty markdown comments on PRs when one or
more json fields contains markdown comments. Because the raw data are
already embedded in the PR-Comment as markdown comment it's a problem
when a json field of such json raw data contains a markdown comment
itself.