fix: fix handling of embedded markdown comments
The action can embedd the raw data of the PR in a PR comment on GitHub.
These raw data is part of the message in form of a markdown comment.
But when the raw data contains a string that itself contains a markdown
comment the generated comment is broken.

This fix is "escaping" the markdown comment Tags '<!--' and '-->' with
a urlencode replacement. With that change the PR comments are rendered
correctly.